:hammer: refactor UserDataPathProvider: use cross-platform APIs
