+++ composeApp/src/commonMain/kotlin/com/crosspaste/path/PlatformUserDataPathProvider.kt
@@ -1,0 +1,8 @@
+package com.crosspaste.path
+
+import okio.Path
+
+interface PlatformUserDataPathProvider {
+
+    fun getUserDefaultStoragePath(): Path
+}
